test: Refactors and simple test
Added super simple test to bump coverage slightly.

Also refactored directories slightly for better naming
