+++ src/components/UserAccessFieldArray/UserAccessFieldArray.js
@@ -30,7 +30,7 @@ import {
 import { NewBox } from '@folio/stripes-erm-components';
 
 import css from './UserAccessFieldArray.css';
-import UserLookupButton from '../UserLookup/UserLookupButton';
+import { UserLookupButton } from '../UserLookup';
 import { useDashboardAccess } from '../../hooks';
 import DashboardAccessInfo from '../DashboardAccessInfo';
 import IfHasAccess from '../IfHasAccess';
