+++ src/components/UserLookup/index.js
@@ -1,1 +1,2 @@
-export { default } from './UserLookup';
+export { default as UserLookupField } from './UserLookupField';
+export { default as UserLookupButton } from './UserLookupButton';
